tests/wpt/tests/css/css-gaps/animation/rule-width-interpolation-repeaters.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="UTF-8">
    <title>rule-width interpolation with repeaters</title>
    <link rel="author" title="Javier Contreras" href="mailto:javiercon@chromium.org">
    <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#column-row-rule-width">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
    <style>
        .parent {
            row-rule-style: solid;
            row-rule-width: 90px;
        }

        .target {
            row-gap: 40px;
            column-gap: 40px;
            row-rule-width: 10px repeat(2, 20px 20px) 20px;
            row-rule-style: solid;
            row-rule-color: black;
            flex-wrap: wrap;
            column-rule-width: repeat(2, 20px 20px);
            column-rule-style: solid;
            column-rule-color: black;
        }
    </style>
</head>
<body>
    <script>
        // Valid Interpolations:
        test_interpolation({
            property: 'row-rule-width',
            from: neutralKeyframe,
            to: '20px repeat(2, 30px 30px) 30px',
        }, [
            { at: -0.3, expect: '7px repeat(2, 17px 17px) 17px' },
            { at: 0, expect: '10px repeat(2, 20px 20px) 20px' },
            { at: 0.3, expect: '13px repeat(2, 23px 23px) 23px' },
            { at: 0.6, expect: '16px repeat(2, 26px 26px) 26px' },
            { at: 1, expect: '20px repeat(2, 30px 30px) 30px' },
            { at: 1.5, expect: '25px repeat(2, 35px 35px) 35px' },
        ]);
        test_interpolation({
            property: 'column-rule-width',
            from: neutralKeyframe,
            to: 'repeat(2, 30px 30px)',
        }, [
            { at: -0.3, expect: 'repeat(2, 17px 17px)' },
            { at: 0, expect: 'repeat(2, 20px 20px)' },
            { at: 0.3, expect: 'repeat(2, 23px 23px)' },
            { at: 0.6, expect: 'repeat(2, 26px 26px)' },
            { at: 1, expect: 'repeat(2, 30px 30px)' },
            { at: 1.5, expect: 'repeat(2, 35px 35px)' },
        ]);
        test_interpolation({
            property: 'column-rule-width',
            from: 'repeat(auto, 20px 20px)',
            to: 'repeat(auto, 30px 30px)',
        }, [
            { at: -0.3, expect: 'repeat(auto, 17px 17px)' },
            { at: 0, expect: 'repeat(auto, 20px 20px)' },
            { at: 0.3, expect: 'repeat(auto, 23px 23px)' },
            { at: 0.6, expect: 'repeat(auto, 26px 26px)' },
            { at: 1, expect: 'repeat(auto, 30px 30px)' },
            { at: 1.5, expect: 'repeat(auto, 35px 35px)' },
        ]);

        // Shouldn't interpolate, length of the list is different:
        test_no_interpolation({
            property: 'row-rule-width',
            from: neutralKeyframe,
            to: 'repeat(2, 30px 30px) 30px',
        });
        test_no_interpolation({
            property: 'row-rule-width',
            from: neutralKeyframe,
            to: 'repeat(2, 30px 30px) repeat(2, 30px 30px)',
        });

        // Shouldn't interpolate, repeaters are not compatible:
        test_no_interpolation({
            property: 'column-rule-width',
            from: neutralKeyframe,
            to: 'repeat(3, 30px 30px)',
        });
        test_no_interpolation({
            property: 'column-rule-width',
            from: neutralKeyframe,
            to: 'repeat(2, 30px 30px 30px)',
        });
        test_no_interpolation({
            property: 'row-rule-width',
            from: neutralKeyframe,
            to: '20px repeat(3, 30px 30px) 30px',
        });
        test_no_interpolation({
            property: 'row-rule-width',
            from: neutralKeyframe,
            to: '20px repeat(2, 30px 30px 30px) 30px',
        });
        test_no_interpolation({
            property: 'column-rule-width',
            from: 'repeat(auto, 20px 20px)',
            to: 'repeat(3, 30px 30px)',
        });
        test_no_interpolation({
            property: 'column-rule-width',
            from: 'repeat(auto, 20px 20px)',
            to: 'repeat(auto, 30px 30px 30px)',
        });

        // Shouldn't interpolate, repeater and a length are not compatible:
        test_no_interpolation({
            property: 'column-rule-width',
            from: neutralKeyframe,
            to: '30px',
        });
    </script>
</body>
